Remove duplicate files include author in filename
